--- src/components/Result/ImageModal.jsx
@@ -1,12 +1,0 @@
-export const ImageModal = () => {
-  return (
-    <section>
-      <h1>
-        이규리의
-        <br />
-        2022년 순간들
-      </h1>
-      <article></article>
-    </section>
-  );
-};
